Suggested (and unfinished) refactoring for the animation system. Changes primarily done to facilitate a more cohesive design: Animations are named collections of frame sequences linked to a sprite sheet. Frame sequences are named collections of frames. Frames define an index, duration and offset within a sprite sheet.
Animations are named globally and can thus be reused in different entities.
Frame sequences are named on the animation itself to avoid name clashes.

Accessors for frames and frame sequences have been added to allow easy manipulation thereof.

In order to make this system actually work with the rest, the following changes still need to be done: The animatable class needs to be updated to use an animation name instead of its own list of animation obejcts as before. The lookup for the animation should happen only when a sprite is rendered to make sure it is not accidentally cached.